backend/requirements.txt
# FastAPI + server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0

# Config & env
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Utils
httpx>=0.27.0
email-validator>=2.0.0
